deepfake-detector-1/requirements.txt
tensorflow
numpy
matplotlib
pandas
opencv-python
scikit-learn
Pillow
